refactor: move client only utils to `client/utils`
